small compiler error

+++ src/test/com/opensymphony/xwork2/config/providers/XmlConfigurationProviderTest.java
@@ -43,11 +43,11 @@ public class XmlConfigurationProviderTest extends ConfigurationTestBase {
         prov.setObjectFactory(container.getInstance(ObjectFactory.class));
         prov.init(configuration);
         List<Document> docs = prov.getDocuments();
-        assertEquals(3, docs.size());
+        assertEquals(3, docs.size() );
 
-        assertEquals(1, XmlHelper.getLoadOrder(docs.get(0)));
-        assertEquals(2, XmlHelper.getLoadOrder(docs.get(1)));
-        assertEquals(3, XmlHelper.getLoadOrder(docs.get(2)));
+        assertEquals(1, XmlHelper.getLoadOrder(docs.get(0)).intValue());
+        assertEquals(2, XmlHelper.getLoadOrder(docs.get(1)).intValue());
+        assertEquals(3, XmlHelper.getLoadOrder(docs.get(2)).intValue());
     }
 
     public void testNeedsReload() throws Exception {
